fix(connector): fix file download
closes #19

+++ src/main/kotlin/com/github/thetric/iliasdownloader/connector/api/model/CourseFile.kt
@@ -10,5 +10,5 @@ data class CourseFile(
     override val url: String,
 
     val modified: LocalDateTime,
-    val size: Int
+    val size: Long
 ) : IliasItem
